refactor: seperate GraphChoice as component
